fix: consolidate FormPopulationCompletedEvent interface definition
- Create shared FormPopulationCompletedEvent interface in fhir-questionnaire-converter types
- Remove duplicate interface definitions from service, controller, and E2E test files
- Update all imports to use the centralized interface definition
- Fix TypeScript re-export issue with proper 'export type' syntax
- Remove unused QuestionnaireOutput import to fix linting error

This addresses the interface duplication issue identified in PR review,
improving type safety and maintainability across the codebase.

All unit tests passing: 34 tests across 3 projects ✅
Interface consistency verified ✅
Linting errors resolved ✅
