+++ apps/form-auto-population-service/src/app/form-population.controller.ts
@@ -1,17 +1,7 @@
 import { Controller, Logger } from '@nestjs/common';
 import { EventPattern, Payload } from '@nestjs/microservices';
 import { FormPopulationService } from './form-population.service';
-
-interface FormPopulationCompletedEvent {
-  formId: string;
-  patientId: string;
-  questionnaireOutput: Array<{
-    question_id: string;
-    question_text: string;
-    answer: string | number | boolean | string[];
-  }>;
-  timestamp: string;
-}
+import { FormPopulationCompletedEvent } from '@form-auto-population/fhir-questionnaire-converter';
 
 @Controller()
 export class FormPopulationController {
